Source code closer can now handle more cases
